pyproject.toml
[project]
name = "img_search"
version = "0.1.0"
description = "Image search engine"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "datasets>=4.1.0",
    "hydra-core>=1.3.2",
    "lmdb>=1.7.3",
    "loguru>=0.7.3",
    "peft>=0.17.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymilvus>=2.6.2",
    "rich>=14.1.0",
    "sentence-transformers>=3.0.0",
    "vllm>=0.10.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.12.11",
]

[tool.pytest.ini_options]
pythonpath = "."
